+++ R/crosstab.R
@@ -27,12 +27,12 @@
 #' crosstab(mtcars$cyl, mtcars$gear, "row")
 #'
 #' # Passing in a data.frame using a pipeline:
-#' library(dplyr) # to access the pipe operator
 #' mtcars %>% crosstab(cyl, gear)
 #' mtcars %>% crosstab(cyl, gear, "row")
 #' 
 #' # This allows for upstream operations
 #' # prior to the crosstab() call:
+#' library(dplyr)
 #' mtcars %>%
 #'   filter(am == 0) %>%
 #'   crosstab(cyl, gear)