+++ R/get_dupes.R
@@ -9,8 +9,7 @@
 #' @export
 #' @examples
 #' get_dupes(mtcars, mpg, hp)
-#' # or called with magrittr pipe %>% :
-#' library(dplyr)
+#' # or called with the magrittr pipe %>% :
 #' mtcars %>% get_dupes(wt)
 #'