+++ R/utils.R
@@ -1,0 +1,17 @@
+# Copied from tidyr/R/utils.R, to export the magrittr pipe
+
+#' Pipe operator
+#'
+#' @description Exported from the magrittr package.  To learn more, run \code{?magrittr::`\%>\%`}.
+#'
+#' @name %>%
+#' @rdname pipe
+#' @keywords internal
+#' @export
+#' @importFrom magrittr %>%
+#' @usage lhs \%>\% rhs
+#' @examples
+#' mtcars %>%
+#'   crosstab(carb, cyl) %>%
+#'   adorn_totals()
+NULL